+++ src/main/java/pl/edu/wszib/prime/numbers/PrimeNumbersApp/service/IMainService.java
@@ -3,5 +3,5 @@ package pl.edu.wszib.prime.numbers.PrimeNumbersApp.service;
 import java.util.List;
 
 public interface IMainService {
-    List<Integer> calculatePrimeNumbers(int num);
+    List<Integer> calculatePrimeNumbers(long num);
 }
